requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0 